Copy refinements into child scope on first use.
MRI does this in vm_cref_new, which I believe is called on first
entry to a method, block, or module body. In their case, they
reference the original overlay module and mark it as shared. Any
subsequent modifications cause it to be replaced on the parent
scopes. In JRuby, I am currently only propagating refinements into
class/module and method bodies, which means blocks and evals and
such still need to search up the hierarchy a bit. I am also doing
a hard copy of the related collection, which could add overhead to
boot time.