Temporary fix for our refined wrappers not actually wrapping.
MRI's logic for an existing method that gets refined expects that
it will be wrapped by a refinement marker. This allows the
original method to be pulled out and look unrefined for other
logic we also mimic. Since currently I do not wrap, but instead
flag, the original method, there's no way to get an original that
isn't refined without duping the method here. This is inefficient
and may lead to peculiar side effects so I will return to using
a full wrapper in a subsequent commit.