Handle refinements within Symbol proc.
This logic modifies all calls with non-literal procs to pass the
current scope through, and modifies the coercion logic to handle
Symbol specially by passing the scope in. This will in turn use
refinements when present.

Note that this is not quite as optimal as the SymbolProc operand
but is still static for the creation of the proc.