Set refine block refinements to parent and search all the way.
This is not quite the right logic, since MRI will copy refinements
into each level and only search the one level. We do not
consistently do that, so we still search the scope hierarchy for
other refinements. This is inefficient and should change, but this
patch at least allows mutually-dependent refine blocks to see
each other.