Time: fix TZ offset
The setTzRelative flag wasn't set everywhere where it should. For one,
the 9.2 changes to initTime() broke Rails' Time.change(). While this
would have been fixable by setting the flag as well, it's just way
too fragile. When a new Time is created based on an instance with
the flag set, it was simply lost. Just returning nil instead of ""
in zone() would fix the basic Rails test, but thing like inspect on
a cloned or otherwise copied-and-adjusted object would still break...

So instead, just remove the flag altogether and evaluate where needed.
This does indeed fix the Rails case, verified this time.