Add --record to enable a full-run Java Flight Recorder recording.
Java Flight Recorder is free as of Java 11 but requires some
arcane JVM command line flags. The most common command-line use of
a profiler in JRuby is roughly equivalent to enabling default
profiling mode in Flight Recorder and just dumping results for the
full run.

Running JRuby with --record will dump, on exit, a .jfr file of the
following format: jruby-flight-record-<PID>.jfr