Rework bash script to avoid second process more.
The following conditions require that the JRuby JVM be launched
as a child process rather than using exec, so that post-execution
logic can run:

* Running with profiling, so that the results can be dumped
* Running under cygwin, which does not support exec

Previously, the following situations also caused us to avoid using
exec:

* Setting any flags that require JRuby to be verified, like -ea
* Setting the NO_BOOTCLASSPATH or VERIFY_JRUBY env vars
* Running on Java 9+, since modules prohibit running in the boot
  classpath

This patch eliminates the cases above where we can still exec.

This fixes a build hang due to MRI's envutil.rb exec_ruby
expecting the child pid to receive a kill signal while cleaning
up after the requested command. For unrelated reasons, the
Thread.stop test in MRI's test_thread.rb test_stop started
hanging, but the shutdown logic in the parent test also hung due
this behavior of the launcher.

This is helped by running the native launcher, which on *nix should
always preserve pid since we can boot and launch the JVM inside
the same process space.

On Windows, the native launcher more frequently will not be able
to keep the JVM in-process, since the launcher we ship is 32-bit
and most Java releases for Windows are 64-bit (so therefore the
executable can't load the JVM dll).
Note that without a native launcher, or with a native launcher on
platforms that don't support exec (e.g. Windows) the PID for the
executed command and the PID for the actual host JVM will still
be different.