Fix spec:ji errors.
It appears JI objects marshalled depends getConstant() to work for marshal.load.
I am guessing we call const_missing to reify the constants (which begs how in
these tests they already exist -- do we not make those constants immediately?).
The work-around with a fixme is to make sure if we are trying to get a path
and that path contains a JI object we will just continue to use getConstant()
for only those segments of the path.