Perform implementer search for aliases with super.
This logic mimics that in MRI's vm_call0_body, which appears to do
something similar to our implementer search when calling an alias,
so that the frame class can reflect the actual source of the alias
at a given call site

We can only detect super use for Ruby methods at the moment, but
that should cover the majority of aliased methods. Core methods
in Java need to have a mechanism for determining if they need the
caller's frame class.

Note also that our implementer search should only fire for
methods defined on a module.