Implement indy trace events and move warnings around.
This allows most trace events in JIT to be always-on. With
always-on trace events we can support all of the `coverage`
library without a flag, and most of the `TracePoint` class, though
the binding provided by set_trace_function and TracePoint#binding
can only be relied upon in debug mode.

The warnings about --debug have been moved to locations where we
would normally provide that binding, so that non-binding uses of
TracePoint and Coverage can execute without an additional flag.

Performance appears to be roughly the same; the additional boolean
checks in interpreter and the indy call sites in JIT do not yet
appear to have a large impact on straight-line execution. However
this is a work in proress and not all tests are passing in
compiled mode, with or without --debug. This may mean there are
trace events not firing that could impact performance more.