Fix -x command line option to skip header
The -x option has been broken since [1], as parseShebangOptions is no
longer run, which is responsible for calling setHasShebangLine. As this
is no longer done, hasShebangLine always returns false, even if there is
a valid shebang.

The fix is to remove the check at this location and move it into the
findScript function, which already searches for the ruby shebang.

[1]: 9e2d6dccb9e481fe25184202ee5a1762bb0fba28

Fix: #4536