Fix encoding/decoding of Trace instructions.
Because :class and :end trace events are now always enabled, we
are emitting them for AOT-compiled Ruby code. However the logic
for encoding/decoding the was not updated after we moved to using
mostly RubySymbol for identifiers in the AST and IR. This commit
updates the logic to properly encode and decode the symbolic name
(if present) associated with the trace event.

See a7d0c0c54d38eb5023e249f5fc96d01fffc6c2a1