Fix thread leaking and incorrect naming in Fiber-based #next.
The original implementation was based on Enumerator code from
Rubinius. However in Rubinius, the Fiber thread does not root
objects any deeper than to the Fiber object itself. In order to
make this not leak threads, all state shared between the Fiber
and the Enumerator is passed in a separate State object.

This also includes minor tweaks, like fixing the to_generator
logic (needed for efficient non-fiber generators), and using the
Fiber.yield method directly as the body of the iteration, so no
intermediate block state is necessary. This last change should
reduce the cost of the iteration loop close to what it was in the
Java-based Enumerator#next.