Eliminate circular static init for Frame and Binding.
Block aggregates a Binding which aggregates a Frame which
aggregates a Block, so having these all initialize statically
caused at least one of those reference to end up null. In #5693
this manifested as a failure to dup the frame for a Symbol proc,
since Symbol procs are initialized with Frame.DUMMY, which static
initialized before Block, so Block.NULL_BLOCK was null and as a
result the block field in Frame was null and triggered asserts.

Block.NULL_BLOCK, as the most heavily used, remains as the sole
static "dummy" in this cycle. NULL_BLOCK is initialized with
a new Binding containing an uninitialized Frame, and then the
Frame is updated to point back to NULL_BLOCK. This avoids the
circulate initialization. Previous references to the dummy Frame
and dummy Binding now use NULL_BLOCK's contents.

See #5692 for the catch-all issue that goes with PR #5693.