RFC: Do not add CLASSPATH to the module path
Treating the CLASSPATH as modules will create automatic modules for all
entries that are not actually modules. This can lead to several problems,
e.g. the automatic module name is determined by the name of the jar or
directory, which fails when one component is not a valid java identifier.
Another issue is that several automatic modules might export the same
package, which is not allowed and leads to an error.

To fix these issues, we can use the module path just for jruby libraries
and pass the CLASSPATH using -classpath, as both can be combined.

TODO:
 - Should $CP also go into the module path?
 - What about the other starters?