Check open status of channel when closing IO.
This allows failed connections from sockets (#5709, #5754) to
raise the actual socket error even though the logic in socket.rb
tries to close the failed socket.

This approach differs from #5713 in that it modifies the final
flush logic to skip closed channels rather than modifying the
IO#closed? logic to do that same check.