Fix lifecycle of UNDEF after autoload fails to define it.
When an autoload runs successfully but does not set the constant
in question, it should clear out any autoload plumbing that was
used but leave the UNDEF value in place, to indicate this value
still is not defined. This change removes the explicit delete of
the constant, to correspond to MRI logic in autoload_reset (which
only sets the actual constant if the autoload set it to a non-null
value) and modifies the getConstantSkipAutoload logic, used by
`defined?` logic, to consider an unresolved undef as undefined if
there's no autoload plumbing installed.