Introduce a wrapper class to eliminate substrings.
Most of the String overhead in the load logic was due to the many
places where a substring is inserted into the index. This change
modifies the collection to aggregate StringWrapper instances,
which share a backing String and just subset a range from it, with
equals and hashCode support. Two pre-allocated wrappers are used
on a thread-local basis to avoid allocating wrappers for index
lookups that are successful.