Back off Suffix abstraction in factor of functional.
Suffix conflated filename construction (basename plus extension)
and library construction, but for Kernel#load we need to treat
files as already having their suffix and create the library based
on the file we find rather than the load string. The Suffix logic
assumed all files would be basename, needing suffix, and having
predetermined logic for creating the library. The new logic only
uses the Suffix abstraction for a require search, which determines
the type of file from the require name.