Replace double-search with a feature check under lock.
This avoids the race between looking up a provided feature and
locking to load that feature. Once the lock is acquired, we check
for the feature again, in case another thread removed its lock and
added the feature while we were preparing to lock and load. This
logic notably does not confirm that the feature is exactly for
the requested file by disassembling its references to feature
paths, since its presence at this point should be enough to know
it has already loaded.