Fall back to a RubyFloat object.
The logic here was attempting to bind a Double to a method handle
that required a RubyFloat, leading to a bad cast. This change
constructs the RubyFloat before binding it.

Note that in this case and in the RubyFixnum case, repeated deopt
of these call sites will keep making new objects to wrap the
numberic value.