Proof-of-concept monomorphic Kernel#send site
This adds a monomorphic cache for Kernel#send calls by saving the
last symbol + method pair retrieved. For cases where this is the
correct method each call, it reduces the overhead of send by
around 50%. In this patch, it is only enabled for non-indy
execution; the indy site will require a similar impl based on
nested call sites.