Provide a JRuby class when checking accessibility.
If we do not pass any class here, the module isOpen check will
only return true if the module is open to "everyone". This is only
the case if the module itself opens up a given package with no
restrictions (i.e. it does not specify a target module).

By passing a class, the isOpen check will use whatever module
JRuby is running as, such as org.jruby.dist in the standard
distribution or org.jruby.core in the jruby-core jar. This in turn
makes it possible to open the module up at the command line and
make non-public methods accessible.

This appears to be the best fix we can do for #5821 in a module
environment. It also appears to fix the issues dispatching to
protected superclass methods from a concrete subclass, also shown
in #5821.