If a class is non-public but open, proceed to bind it.
The existing logic will fail to define any Ruby proxy class +
method stubs when a given class is nonpublic and a generic
setAccessible (unrelated to the class) would fail. This makes it
impossible to allow JRuby to bind the class when it has been
explicitly opened up by e.g. `add-opens` flags.

This patch uses the module's open status to determine whether the
package has been opened up to JRuby. If it has, we proceed to bind
the class as if it were possible to set it accessible.