Add --environment option.
This flag gathers relevant environment info from environment
variables, JVM argument files, and command line flags to produce
a log of every config used to boot the JVM for JRuby.

Note this does not currently include .jrubyrc contents because
that file is processed after JVM startup and is only used for
JRuby system properties. Support could be added but would
duplicate the logic we do at runtime (unless we also move that
logic to the launchers).