Wrap Process.wait* calls with pthread_kill logic.
This addresses #1050 in the same way CRuby implements this. In
CRuby, the wait functions are called directly, and the
`pthread_kill` function is used to force the call to be
interrupted. It turns out the JVM also uses `pthread_kill` to
interrupt blocking native calls, so we use the same binding they
do.

Note that this will only be used on non-Windows, but the JDK's
binding of `NativeThread.signal` might already be stubbed out
there.