Experiment in nuking the JIT counter in methods if the threshold has been reached BUT it has taken longer than -Xjit.time.delta ns since the last time saved when the counter was set to 0.
This is totally untuned in that I made up a number and it I suspect is too
dramatically not JITting methods.  In running an oj bench and gem list (with
and without --dev) and starting up rails the times do not appear to be much
different.

After tuning the value (which I guess would be for this machine -- a follow
up to this work will be whether one magic constant can work here) which would
be figuring out proper time delta and also determing whether any important
methods which should have JITted had not there is a followup change I want to
try: change threshold to be more than just a method call counter.  I would
like to also add threadpolls as an increment value.  Then larger methods which
are not called as much will end up as JIT candidates.