Cache most direct method handles.
A large part of the cost of building a MethodHandle chain appears
to be the final direct method handle. This commit caches most of
the DMH that we were recreating repeatedly on hot path code. This
appears to help startup time a measurable amount, with `rails -h`
going from 5.2s to 4.6s on my system.