Eliminate race in indy-based //o caching.
This logic incorrectly (but atomically!) assigned the current
value of `cache` back to itself, which if cache were null would
simply reassign null. It then proceeded to correctly re-assign
the call site's target to the actual new Regexp object, but
because the CAS was broken this might happen more than once. This
in turn broke code that depended on //o always returning the first
object it produced.

I've flipped the logic and cleaned up the atomic update:

* If the regexp is not /o, just return it. Otherwise...
* Attempt to CAS the cache field. This will succeed exactly once.
* If the update succeeds, permanently bind the call site to the
  value we just set and return that value.
* If the update fails, then the cache has been assigned on another
  thread; reacquire it and return that result.