Check for jitted method after tryJit.
In order to make jit.threshold=0 immediately switch to the newly
jitted method, we need to check actualMethod after we tryJit. The
previous code allowed the calling thread to proceed to the
interpreted version, giving it at least one chance to run. This in
turn led to exposing a known race in //o caching when switching
from interpreter to JIT because the JIT would run first, then the
interpreter would cache //o, but the JIT already compiled a second
bit of evaluation and caching logic. Normally, if the interprter
executes before the JIT, the JIT will reuse the //o result from
the interpreter.

This patch restores order to tryJit and the actualMethod check.