Allow blocks that happen within an eval to JIT.
Two things prevent a block in an eval from jitting:

* If the method containing the eval has not been jitted, when
  we walk up to it we will jit that body and the eval will just
  be seen as a method call. No further compilation will walk into
  the eval.
* If the method containing the eval has already been jitted, we
  similarly walk up to that method, see it has been jitted, and
  proceed no further.

This PR is a small proof-of-concept that changes how the parent
walking happens. Instead of walking all the way to the method
body, this modified logic will walk only up to the nearest
unjitted body, which in this case would be the eval. The eval
body itself won't actually jit, but in order to allow the block it
contains to jit we must at least process it to full IR. Then the
block can also JIT and we proceed from there with the jitted
version.

This still has the problem of walking up to a nearest method body
that has not been jitted; we will stop there and compile the eval
as a call.

I am submitting this as a partially working PR for us to debate
whether this is likely to be valuable. The only cases that do not
jit now would be a block directly in an eval body, like so:

```ruby
def foo
  eval "
    proc { woohoo }
  "
end
```

Methods defined within an eval act as their own JIT root, so they
are not affected by this issue. I have not done any research to
see how common it is to eval a proc for later (heavy) use.