Continue trying to acquire the lock after interrupt.
The original code would immediately error out if a contended lock
acquisition were interrupted before it could acquire the lock.
This led to a behavioral difference from MRI, which does not raise
errors. It is not clear whether that's because locks in MRI can
never be truly contended (if the lock is not available the thread
is put to sleep and woken by the lock owner) or because they
explicitly try again to acquire the lock after waking.

In this version of the code, we opt to continue trying to acquire
the lock via a loop, unless a thread event like raise or kill tell
us to divert from the normal path of execution. The only possible
outcomes from interrupting a lock acquisition are to keep trying
until we have the lock or error out. This patch opts for the
former.

Fixes #5875.