Restructure indy-based math binding. (#5939)
This primarily adds a hard RubyFixnum or RubyFloat type check on
self, to ensure we're not even attempting fast bindings for other
types. It also reduces overall binding overhead in that case by
moving direct indy logic out of the fail path and moves a couple
more adapted method handles to constants.

Fixes #5938.