Switch warn to use checkType again to match MRI.  Attempt at fixing checkType to notice cases where it extends a primal type in Ruby (e.g. String).
I believe the risk to this is very tiny. No existing uses which worked before
will fail from this.  The extra permissiveness will allow core extended types
which are identified as T_* in MRI to continue to work here because we can
check that relationship via inheritance check (since an extended String
will end up having everything a String should have).

Final note to ponder...why does CheckType even exist in MRI?  It is seemingly
more regressive than CheckXXXType as it will not perform protocol coercion.
In some cases perhaps no methods calls can be allowed to happen but in auditing
how it is used in MRI I do not see any case like that.