Handle String#sub/#gsub with string arg like CRuby
`String#sub` and `#gsub` called with a String arg should use string indexing
and not compile a Regexp.

This requires:

* Constructing MatchData instances appropriately so that `$~` et al get
populated correctly (but only in those situations where it would
happen).

* Using the `Regexp` static method `regsub()` to perform the replacement
string parsing so that special sequences (`\0`, `\'`, etc) are correctly
expanded and inserted.

@enebo wrote a new `StringSupport.index()` method that makes the
position indexing easier, which we use here.

As a style refactor, the `gsubCommon19()` methods were renamed
`gsubCommon()`, since there's no 1.9 distinction anymore and the methods
are all private.