Cache method_missing target.
This logic will cache the target class's method_missing impl when
a method call fails to find a natural method table entry. This
scenario was a source of failed call sites in Redmine's json
builder and in the Rails StringInquirer class used for Rails.env.
With this change in place Redmine json issue requests improved by
roughly 15%.