Reduce field traversals converting booleans
This commit attempts to reduce the amount of field traversal done
to convert a primitive boolean value into a RubyBoolean. Typically
this conversion was done by calling runtime.newBoolean(val) or
RubyBoolean.newBoolean(runtime, val) but in many cases this call
required retrieval of the runtime from either the context or the
current object's metaclass. In cases where the context is already
available and the runtime is not needed, I have changed to using
RubyBoolean.newBoolean(context, val) to avoid the runtime field
traversal. In addition this commit changes any runtime.getTrue()
and runtime.getFalse() calls to context.tru and context.fals when
the context is already present and the runtime is not otherwise
needed.

Any calls to the runtime versions of these methods where the
runtime was already available or needed for other purposes have
been left alone to reduce the size of the diff.