Reduce the chance of lazy enumerator leaks.
Lazy enumerators still may leak in the presence of Enumerator#next
calls, but this at least isolates the instantiating Lazy from the
newly-created Lazy, avoiding a backreference that might keep the
original object alive.

Given a test as shown in #6004 that spins up many Lazy instances,
this patch appears to work well. It did not appear to fix the
reported leaks in that issue, however.

Example code that leaked before but is stable with this patch:

```ruby
while true
  [1,2,3].lazy.reject {|x| x < 3 }.next
  JRuby.gc
end
```