Ensure vacated wrappers won't raise NPE if inspected
We will frequently have vacated wrappers in this cache, because
we wish to use them for subsequent lookups without constructing
anew. Normally this would not cause an error; the wrappers are
populated with a string before being used. However in #6039 the
aggressive thread-local leak detector from Tomcat stumbles over
these empty wrappers, triggering an NPE while attempting to get a
hash code. This commit avoids the Tomcat issue by short-circuiting
hashCode and equals to return 0 and false respectively when the
wrapper has a null string.

Fixes #6039.