Fix lazy instruction loading with AOT/ir-serialization.
The main fix was that in IRReader.load we would set up a callback to make
instructions lazy but then in third loop going over the newly made scopes
we would call computeScopeFlags which would immediate invoke the callback
to retrieve the instructions.

Those three loops are now a single loop.

Fixing that issue revealed a second issue in that getting instructions
from a lazy callback would happen after some predicate methods occurred
in interpretercontext.  Those predicates were based on flags which were not
calculated because we did not read the instructions yet.  Guards in those
methods to retrieve instructions ended up fixing that second issue.