Lazily set up closure scopes using INDY MAGIC
This needs a little explanation...

When scopes were set before executing jit code we could just
retrieve the already-created closure scope from a static field.

With scopes now being created on the fly from a descriptor, we
need to be able to construct those scopes lazily. The logic now
does the following:

* Retrieve the scope field as before (via method handle)
* If null, retrieve the parent scope (via method handle) and
  restore the closure scope using that and the closure's scope
  descriptor
* Set the resulting scope back into the closure's scope field

This involves a lot of indy magic (passing a "putstatic" handle
for the closure scope and a "getstatic" handle for the parent
scope) but it works and once bootstrapped none of this happens
again.

Note that for top-level closures, this still depends on the script
body scope having been initialized already. Since at the moment
of this commit we do not restore the script scope from jit code,
it must be set for this to work properly at the top level. Soon
the script scope will also be restored by jit code and the
external set will not be necessary.