Fix wrapping logic by passing self through
Only the interpreter actually passed through the wrapper version
of the top self. Previous commits broke that path by moving the
top self clone into a utility method and not passing it on, based
on the logic of the other paths. This commit restores the clone
logic and modifies all paths to pass it through. This also gets
wrapped loading working for AOT mode, which was broken in the same
way before.