Be more explicit in not optimizing in the presence of END blocks.
Up until 10ec7bef9fb0a910c14b8583e6bdd25ed521919d
TestParser.testLineNumber was passing BUT before and after this commit it
was triggering an assert thinking there should be no nested closures.  The
END block was registered so they both would throw.  What is weird is before
the commit above the test would claim it passed????

In truth, this optimization need not run on a scope with an END block so
disabling.